re-organize .h and .cpp
