+++ roboteq_controller/include/roboteq_controller/roboteq_controller_node.h
@@ -1,109 +1,92 @@
-#ifndef ROBOTEQ_MOTOR_CONTROLLER_DRIVER_MAIN_H
-#define ROBOTEQ_MOTOR_CONTROLLER_DRIVER_MAIN_H
+#pragma once
+
+#include <boost/lexical_cast.hpp>
+#include <boost/algorithm/string/split.hpp>
+#include <boost/algorithm/string/trim.hpp>
+#include <boost/algorithm/string/replace.hpp>
+#include <boost/algorithm/string/classification.hpp>
+
+#include <serial/serial.h>
+#include <iostream>
+#include <sstream>
+#include <typeinfo>
 
 #include <ros/ros.h>
-#include <tf/tf.h>
 
-//! ROS standard msgs
-// #include <geometry_msgs/Quaternion.h>
-// #include <geometry_msgs/Vector3Stamped.h>
-// #include <sensor_msgs/Imu.h>
-// #include <sensor_msgs/NavSatFix.h>
-// #include <sensor_msgs/Joy.h>
-// #include <sensor_msgs/TimeReference.h>
-// #include <sensor_msgs/BatteryState.h>
-// #include <sensor_msgs/Image.h>
-#include <geometry_msgs/TransformStamped.h>
+#include <std_msgs/String.h>
+#include <std_msgs/Empty.h>
 #include <geometry_msgs/Twist.h>
 #include <nav_msgs/Odometry.h>
 
-// #include <std_msgs/UInt8.h>
-// #include <std_msgs/Int16.h>
-// #include <std_msgs/Float32.h>
-// #include <std_msgs/String.h>
+#include "roboteq_controller/querylist.h"
+#include "roboteq_controller/channel_values.h"
+#include "roboteq_controller/config_srv.h"
+#include "roboteq_controller/command_srv.h"
+#include "roboteq_controller/maintenance_srv.h"
 
-#include <tf/transform_broadcaster.h>
 
-#include <roboteq_controller/channel_values.h>
-
-#include <roboteq_controller/config_srv.h>
-#include <roboteq_controller/command_srv.h>
-#include <roboteq_controller/maintenance_srv.h>
-
-#include <serial/serial.h>
-#include <roboteq_controller/querylist.h>
-#include <boost/algorithm/string/regex.hpp>
-#include <boost/regex.hpp>
-
-namespace roboteq{
-
-class Driver
+class RoboteqDriver
 {
-public: 
-	//Driver(ros::NodeHandle& nh);
-	//Driver();
-	//~Driver();
-	ros::Subscriber cmd_vel_sub;
-	ros::Publisher read_publisher;
+public:
+	RoboteqDriver(ros::NodeHandle, ros::NodeHandle);
+	~RoboteqDriver(){
+		if (ser_.isOpen()){
+			ser_.close();
+		}
+	}
+
+private:
+	ros::NodeHandle 		nh_, nh_priv_;
+
+	ros::ServiceServer 		configsrv_;
+	ros::ServiceServer 		commandsrv_;
+	ros::ServiceServer 		maintenancesrv_;
 	
-	ros::ServiceServer configsrv;
-	ros::ServiceServer commandsrv;
-	ros::ServiceServer maintenancesrv;	
-	//ros::ServiceClient configsrv_client;
-	//void diff_drive(int speed1, int speed2, float wheel_rad, float wheel_dist, int encoder_coef);
-	//void cmd_vel_callback(sensor_msgs::Joy::ConstPtr& msg);
-	void connect();
-	void run();
-	void roboteq_subscriber();
-	void roboteq_publisher();
-	void cmd_vel_callback(const geometry_msgs::Twist& msg);
-	//void initSub();
-	
-	
+	// Serial
+	serial::Serial 			ser_;
+	std::string 			serial_port_;
+	int32_t 				baudrate_;
+
+	// Pub & Sub
+	ros::Subscriber 		cmd_vel_sub_;
+	ros::Publisher 			serial_read_pub_;
+	std::vector<ros::Publisher> publisherVecH_, publisherVecL_, publisherVecG_;
+
+	ros::Timer timerH_, timerL_, timerG_;
+
+
+	bool 					closed_loop_,
+							diff_drive_mode_;
+
+	double 					wheel_circumference_,
+							track_width_,
+							max_rpm_;
+
+	std::string 			cmd_vel_topic_;
+
 	int channel_number_1;
 	int channel_number_2;
-	int frequencyH;
-	int frequencyL;
-	int frequencyG;
-	
-	
-	
-	
-	void roboteq_services();
-	bool configservice(roboteq_controller::config_srv::Request& req,     	roboteq_controller::config_srv::Response& res);
-	
-	bool commandservice(roboteq_controller::command_srv::Request& req,     	roboteq_controller::command_srv::Response& res);
-	
-	bool maintenanceservice(roboteq_controller::maintenance_srv::Request& req,     	roboteq_controller::maintenance_srv::Response& res);
-	
-	
-private:
-	int baud_rate;
-	std::string port;
-	std::string firmware;
-	int channel;
-	
-	
-	geometry_msgs::TransformStamped tf_msg;
-	tf::TransformBroadcaster odom_broadcaster;
-	nav_msgs::Odometry odom_msg;
-    nav_msgs::Odometry odom;
+	int frequencyH_;
+	int frequencyL_;
+	int frequencyG_;
 
-	enum fault_flag
-	{
-	NO_FAULT = 0,
-	OVERHEAT = 1,
-	OVERVOLTAGE = 2,
-	UNDERVOLTAGE = 4,
-	SHORT_CIRCUIT = 8,
-	EMERGENCY_STOP = 16,
-	SETUP_FAULT = 32,
-	MOSFET_FAILURE = 64,
-	STARTUP_CONFIG_FAULT = 128,
-	};
-	
-		
-};	
-}	
-#endif // ROBOTEQ_MOTOR_CONTROLLER_DRIVER_MAIN_H
-	
+
+	void cmdSetup();
+	void cmdVelCallback(const geometry_msgs::Twist &);
+	void powerCmdCallback(const geometry_msgs::Twist &);
+	bool configService(roboteq_controller::config_srv::Request &, roboteq_controller::config_srv::Response &);
+	bool commandService(roboteq_controller::command_srv::Request &, roboteq_controller::command_srv::Response &);
+	bool maintenanceService(roboteq_controller::maintenance_srv::Request &, roboteq_controller::maintenance_srv::Response &);
+	void initializeServices();
+	void run();
+
+	void formQuery(std::string, 
+				std::map<std::string,std::string> &, 
+				std::vector<ros::Publisher> &,
+				std::stringstream &);
+
+	void highFreqCallback	(const ros::TimerEvent &);
+	void lowFreqCallback	(const ros::TimerEvent &);
+	void generalFreqCallback(const ros::TimerEvent &);
+
+};
